fix: use long type for cache token fields in ES schema and migration
The field was already mapped as long in ES (from OTEL intValue ingestion),
so the migration must use long to avoid illegal_argument_exception.

+++ langwatch/elastic/migrations/202602082300_add_cache_token_metrics.ts
@@ -13,8 +13,8 @@ export const migrate = async (_migrationKey: string, client: ElasticClient) => {
     properties: {
       metrics: {
         properties: {
-          cache_read_input_tokens: { type: "integer" },
-          cache_creation_input_tokens: { type: "integer" },
+          cache_read_input_tokens: { type: "long" },
+          cache_creation_input_tokens: { type: "long" },
         },
       },
       spans: {
@@ -22,8 +22,8 @@ export const migrate = async (_migrationKey: string, client: ElasticClient) => {
         properties: {
           metrics: {
             properties: {
-              cache_read_input_tokens: { type: "integer" },
-              cache_creation_input_tokens: { type: "integer" },
+              cache_read_input_tokens: { type: "long" },
+              cache_creation_input_tokens: { type: "long" },
             },
           },
         },
